Fix expiration date conversion when the day is zero
Many manufacturers inform just the month and the year of the expiration date in their barcodes. If that happens, the day usually comes as zero. In such scenario, the date parsing code ends up converting the expiration date to the last day of the previous month. E.g., a barcode contaning 190700 in the date position becomes Jun 30, 2019. It should be converted to Jul 31, 2019, instead.