Deprecation notice
I've been trying to contribute to this project for some time. I created a few pull requests a few months ago. Unfortunately, it looks like this repository is inactive and the author is not accepting pull requests anymore. For this reason, I've created a fork to do the required maintenance.

Please, use the [new repository](https://github.com/hprange/BarcodeParser) if you want to contribute pull requests or fill issues.